+++ node/directives/checkUserAccess.ts
@@ -25,8 +25,12 @@ export class CheckUserAccess extends SchemaDirectiveVisitor {
         vtex: { adminUserAuthToken, storeUserAuthToken, logger },
       } = context
 
-      const { hasAdminToken, hasValidAdminToken, hasCurrentValidAdminToken } =
-        await validateAdminToken(context, adminUserAuthToken as string)
+      const {
+        hasAdminToken,
+        hasValidAdminToken,
+        hasCurrentValidAdminToken,
+        hasValidAdminTokenFromStore,
+      } = await validateAdminToken(context, adminUserAuthToken as string)
 
       const {
         hasAdminTokenOnHeader,
@@ -34,7 +38,8 @@ export class CheckUserAccess extends SchemaDirectiveVisitor {
         hasCurrentValidAdminTokenOnHeader,
       } = await validateAdminTokenOnHeader(context)
 
-      const { hasApiToken, hasValidApiToken } = await validateApiToken(context)
+      const { hasApiToken, hasValidApiToken, hasValidApiTokenFromStore } =
+        await validateApiToken(context)
 
       const { hasStoreToken, hasValidStoreToken, hasCurrentValidStoreToken } =
         await validateStoreToken(context, storeUserAuthToken as string)
@@ -62,6 +67,8 @@ export class CheckUserAccess extends SchemaDirectiveVisitor {
           hasValidStoreToken,
           hasAdminTokenOnHeader,
           hasValidAdminTokenOnHeader,
+          hasValidAdminTokenFromStore,
+          hasValidApiTokenFromStore,
         },
         'CheckUserAccessAudit'
       )
@@ -87,6 +94,8 @@ export class CheckUserAccess extends SchemaDirectiveVisitor {
           hasStoreToken,
           hasAdminTokenOnHeader,
           hasValidAdminTokenOnHeader,
+          hasValidAdminTokenFromStore,
+          hasValidApiTokenFromStore,
         })
         throw new AuthenticationError('No token was provided')
       }
@@ -111,6 +120,8 @@ export class CheckUserAccess extends SchemaDirectiveVisitor {
           hasValidStoreToken,
           hasAdminTokenOnHeader,
           hasValidAdminTokenOnHeader,
+          hasValidAdminTokenFromStore,
+          hasValidApiTokenFromStore,
         })
         throw new ForbiddenError('Unauthorized Access')
       }